# Conflicts:
#	.idea/caches/build_file_checksums.ser
#	app/src/main/java/mx/qsistemas/infracciones/db_web/entities/InfringementPayorder.kt
#	app/src/main/java/mx/qsistemas/infracciones/net/NetworkApi.kt

+++ app/src/main/java/mx/qsistemas/infracciones/net/request_web/PaymentRequest.kt
@@ -2,14 +2,14 @@ package mx.qsistemas.infracciones.net.request_web
 
 import com.google.gson.annotations.SerializedName
 
-data class PaymentRequest(@SerializedName("discount") val discount: Double,
+data class PaymentRequest(@SerializedName("discount") val discount: Float,
                           @SerializedName("folio_payment") val folio_payment: String,
                           @SerializedName("observations") val observations: String,
                           @SerializedName("payment_date") val payment_date: String,
                           @SerializedName("payment_method") val payment_method: String = "CARD",
-                          @SerializedName("rounding") val rounding: Double,
-                          @SerializedName("subtotal") val subtotal: Int,
-                          @SerializedName("surcharges") val surcharges: Int,
+                          @SerializedName("rounding") val rounding: Float,
+                          @SerializedName("subtotal") val subtotal: Float,
+                          @SerializedName("surcharges") val surcharges: Float,
                           @SerializedName("token") val token: String,
-                          @SerializedName("total") val total: Int,
+                          @SerializedName("total") val total: Float,
                           @SerializedName("num_authorization") val num_authorization: String)
